[Fix #7] Move node_pattern.rb, processed_source_spec.rb and rubocop/token_spec.rb into ast/
Done in two operations to facilitate git-blame
