+++ lib/rubocop/ast/node_pattern.rb
