+++ lib/rubocop/ast/processed_source.rb
