+++ lib/rubocop/ast/token.rb
