+++ spec/rubocop/ast/node_pattern_spec.rb
