+++ spec/rubocop/ast/token_spec.rb
